<Adding PWM Driver Module and Testbench to Motherboard Framework>
<PWM_Driver.v takes a 16-bit data bus as input as well as a reset signal and a load signal. The top 8 bits of the data
bus signify the duty cycle, while the lower 8 bits signify the period. The duty cycle is input as a percentage of 255.
For example, should the user require a 20% duty cycle, 'b51 should be loaded since 0.2 * 255 = 51.

The value loaded to the period is passed through an internal clock divider to slow the system's default 10ns clock to the
desired period. Internally, 2 is added to the value loaded by the user to both slow the maximum period to approx. 10us
and to increase the resolution at shorter periods. The resulting signal from the clock generator is used to increment
the counters that generate a PWM signal, which count to a total of 255 (the width of the duty cycle inpout). Therefore,
the resulting frequency will equate to 100_000_000/(((INPUT+2)/2)*1020)Hz. The smallest period attainable with this
method is 10.2us with an inpout of 'b0, while the largest is 657.9us with an input of 'b255.

Test_PWMDriver.v tests 25 PWM signals simultaneously for an arbitrary amount of time. 5 different duty cycles and 5
different periods are tested in every combination possible. The input values are identical: 0, 64, 128, 191, and 255
are tested for both duty cycle and period. These numbers approximate 0%, 25%, 50%, 75%, and 100% of 255, respectively,
to provide a diverse series of waveforms which can then be analyzed to determine accuracy to desired results.>